+++ src/settings.ts
@@ -10,8 +10,8 @@ import {
   TFolder,
 } from 'obsidian';
 
-import MobilePlugin, { CommandManager } from './main';
 import { GestureCommand } from './gesture-handler';
+import MobilePlugin from './main';
 
 // Type for Obsidian's internal commands API (not in public API)
 
@@ -335,13 +335,13 @@ export class CommandSuggestModal extends FuzzySuggestModal<Command> {
   onSubmit: (result: Command) => void;
   commands: Command[];
 
-  constructor(app: App, onSubmit: (result: Command) => void) {
+  constructor(
+    public app: App,
+    onSubmit: (result: Command) => void,
+  ) {
     super(app);
     this.onSubmit = onSubmit;
-    this.commands = Object.values(this.commandManager?.commands || {});
-  }
-  get commandManager(): CommandManager | undefined {
-    return (this.app as { commands?: CommandManager }).commands;
+    this.commands = Object.values(app.commands?.commands || {});
   }
 
   getItems(): Command[] {
@@ -858,7 +858,7 @@ export class ToolbarEditor extends Modal {
       );
 
     this.toolbar.commands.forEach((cmdId, index) => {
-      const command = this.commandManager?.findCommand?.(cmdId);
+      const command = this.app.commands?.findCommand?.(cmdId);
 
       const setting = new Setting(container)
         .setName(command?.name || cmdId)
@@ -980,9 +980,6 @@ export class ToolbarEditor extends Modal {
     );
     return this;
     // Additional settings for editing commands can be added here
-  }
-  get commandManager(): CommandManager | undefined {
-    return (this.app as { commands?: CommandManager }).commands;
   }
 
   onClose() {