+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the Mobile UX plugin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.2] - 2025-12-21
+
+### Changed
+
+- **Search**: Improved selection command bar layout and visibility.
+
+### Fixed
+
+- **Toolbar**: Fixed text hidden under the toolbar.
+
 ## [1.4.1] - 2025-12-20
 
 ### Fixed